+++ CHANGELOG.md
@@ -1,1 +1,3 @@
+## Change Log
 
+- [hifumi.py](https://github.com/hifumibot/hifumibot/blob/3a87be1970b4e3a3b5986c5773835a6e03954836/bot/hifumi.py) has been refactored.